[ghstack-poisoned]

+++ torch/csrc/jit/runtime/argument_spec.h
@@ -241,7 +241,7 @@ struct CompleteArgumentInfo;
 struct CompleteArgumentSpec {
   CompleteArgumentSpec(bool with_grad, at::ArrayRef<IValue> inputs)
       : ninputs(inputs.size()) {
-    int32_t all_dims = 0;
+    int64_t all_dims = 0;
     const auto num_inputs = inputs.size();
     for (const auto i : c10::irange(num_inputs)) {
       if (!inputs[i].isTensor())
